requirements.txt
absl-py==0.13.0
astunparse==1.6.3
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625814692085/work
chardet==3.0.4
charset-normalizer==2.0.3
click==6.7
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262147379/work
configparser==5.0.2
cryptography @ file:///tmp/build/80754af9/cryptography_1616767007030/work
cycler==0.10.0
docker-pycreds==0.4.0
future @ file:///tmp/build/80754af9/future_1607571303524/work
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.33.1
google-auth-oauthlib==0.4.4
grab-screen==0.0.13
grpcio==1.39.0
idna==2.6
iso8601==0.1.16
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mss==3.0.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264704411/work
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
pafy==0.5.5
pandas==1.3.1
pathtools==0.1.2
Pillow @ file:///tmp/build/80754af9/pillow_1625655408865/work
pip-autoremove==0.9.1
promise==2.3
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyperclip==1.5.27
pyserial==3.5
pyshark==0.4.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
scipy==1.7.0
seaborn==0.11.1
sentry-sdk==1.3.0
shortuuid==1.0.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
torch==1.9.0
torchaudio==0.9.0a0+33b2469
torchvision==0.10.0
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3==1.22
wandb==0.11.0
Werkzeug==2.0.1
youtube-dl==2021.6.6
